changed lat/long to double instead of float. MySql doesnt allow enough decimal places

+++ app/src/main/java/com/example/ger/myapplication/CreateNoteAsyncTask.java
@@ -25,8 +25,8 @@ class CreateNoteAsyncTask extends AsyncTask<String, Void, Note> {
 
         note = new Note( );
         note.setText(text);
-        note.setLatitude((float)latLng.latitude);
-        note.setLongitude((float)latLng.longitude);
+        note.setLatitude(latLng.latitude);
+        note.setLongitude(latLng.longitude);
 
     }
