+++ app/src/main/java/com/example/ger/myapplication/Note.java
@@ -6,9 +6,9 @@ public class Note {
 
     private String text;
 
-    private float latitude;
+    private double latitude;
 
-    private float longitude;
+    private double longitude;
 
     public int getId( ) {
         return id;
@@ -26,19 +26,19 @@ public class Note {
         this.text = text;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
